Fix the OCR findings on M9, M14 and N1
- ceres.h declared sys_exit, sys_exit_status and sys_reset a second
  time without noreturn; both declarations carry it now.
- test_sort squared its values in int, which overflows for values
  above 46340; the square is taken in unsigned arithmetic.
- stdbit.h's trailing-zero count is a plain zero test instead of a
  min with the width, like its leading-zero count; stdbit.h and
  stdckdint.h assert the 32-bit long their _ul bounds assume.

Dismissed: setjmp without returns_twice - ceresc does not know the
attribute (it would be W2002, an error under -Werror), and its setjmp
contract is verified in Ceres-C and tested at -O2. _Alignas as a
macro - ceresc does not predefine it, and the header already says
alignas takes a number, not a type.

+++ tests/test_sort.c
@@ -47,8 +47,8 @@ static int same_values(const int* a, const int* b, int n)
     unsigned int s1 = 0, s2 = 0, q1 = 0, q2 = 0;
     for (int i = 0; i < n; i++)
     {
-        s1 += (unsigned int)a[i]; q1 += (unsigned int)(a[i] * a[i]);
-        s2 += (unsigned int)b[i]; q2 += (unsigned int)(b[i] * b[i]);
+        s1 += (unsigned int)a[i]; q1 += (unsigned int)a[i] * (unsigned int)a[i];
+        s2 += (unsigned int)b[i]; q2 += (unsigned int)b[i] * (unsigned int)b[i];
     }
     return s1 == s2 && q1 == q2;
 }
